requirements.txt
pydicom
pyinstaller
Numpy
Pillow
matplotlib